% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-retweets.R
\name{getRetweets}
\alias{getRetweets}
\title{Returns the list of user IDs that retweeted a specific tweet.}
\usage{
getRetweets(id = NULL, oauth_folder, cursor = -1, verbose = TRUE,
  sleep = 1)
}
\arguments{
\item{id}{The numerical ID of the desired status}

\item{oauth_folder}{folder where OAuth tokens are stored.}

\item{cursor}{See \url{https://dev.twitter.com/docs/api/1.1/get/statuses/retweeters/ids}}

\item{verbose}{If \code{TRUE}, prints information about API calls on console}

\item{sleep}{Number of seconds to sleep between API calls.}
}
\description{
\code{getRetweets} connects to the REST API of Twitter and returns a
list of up to 100 user IDs belonging to users who have retweeted the
tweet specified by the id parameter.
}
\examples{
\dontrun{
## Download list of recent user IDs retweeting a tweet by Hillary Clinton
 rts <- getRetweets(id='653733796408377344', oauth_folder="~/credentials/twitter")
}
}
\author{
Pablo Barbera \email{pablo.barbera@nyu.edu}
}

