% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/search-users.R
\name{searchUsers}
\alias{searchUsers}
\title{Returns list of users related to a search query}
\usage{
searchUsers(q = NULL, count = 100, oauth_folder, verbose = TRUE)
}
\arguments{
\item{q}{The search query to run against people search}

\item{count}{Number of potential user results to retrieve}

\item{oauth_folder}{folder where OAuth tokens are stored. It can also be the name
of a file that contains the token, or a csv file with the format: consumer_key,
consumer_secret,access_token,access_token_secret.}

\item{verbose}{shows additional ouput about token usage in console}
}
\description{
\code{searchUsers} Provides a simple, relevance-based search interface
to public user accounts on Twitter. Try querying by topical interest,
full name, company name, location, or other criteria. Exact match searches
are not supported. Only the first 1,000 matches are available.
}
\examples{
\dontrun{
## Search users using query "pablo"
 userdata <- searchUsers(q="pablo", count=200, oauth_folder="~/Dropbox/credentials")
}
}
\author{
Pablo Barbera \email{pablo.barbera@nyu.edu}
}

