\docType{package}
\name{smapp-package}
\alias{smapp}
\alias{smapp-package}
\title{Tools for analysis of Twitter data}
\description{
  This package provides a series of functions that allow
  lab members to access the Mongo database of tweets in the
  lab computer and easily compute summary statistics
}
\author{
  Pablo Barbera \email{pablo.barbera@nyu.edu}
}
\seealso{
  \code{\link{count.tweets}}, \code{\link{extract.tweets}},
  \code{\link{extract.retweets}},
  \code{\link{extract.hashtags}},
}

