% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/smappR-package.R
\docType{package}
\name{smappR-package}
\alias{smappR}
\alias{smappR-package}
\title{Tools for analysis of Twitter data}
\description{
This package provides a series of functions that allow lab
members to access the Mongo database of tweets in the lab computer
and easily compute summary statistics
}
\author{
Pablo Barbera \email{pablo.barbera@nyu.edu}
}
\seealso{
\code{\link{count.tweets}}, \code{\link{extract.tweets}},
\code{\link{extract.retweets}}, \code{\link{extract.hashtags}},
}

