% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/traceplot-ideology.R
\name{traceplot.ideology}
\alias{traceplot.ideology}
\title{Displays trace plot for MCMC chains of ideology estimator}
\usage{
traceplot.ideology(results, par = "theta")
}
\arguments{
\item{results}{ideology estimation object from \code{estimate.ideology} function}

\item{par}{parameter for which trace plot is to be displayed}
}
\description{
Displays trace plot for MCMC chains of ideology estimator
}
\examples{
\dontrun{
## download list of friends for a given user
 friends <- getFriends(screen_name = "p_barbera", oauth_folder="oauth")
## estimating ideology
 results <- estimate.ideology(friends)
## trace plot
 traceplot.ideology(results, "theta")
}
}
\author{
Pablo Barbera \email{pablo.barbera@nyu.edu}
}

