% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tweets-to-df.R
\name{tweetsToDF}
\alias{tweetsToDF}
\title{Converts object of class "tweets" to data frame}
\usage{
tweetsToDF(tweets)
}
\arguments{
\item{tweets}{object of class tweets}
}
\description{
Converts object of class "tweets" to data frame
}
\author{
Pablo Barbera \email{pablo.barbera@nyu.edu}
}

