% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tweets-to-mongo.R
\name{tweetsToMongo}
\alias{tweetsToMongo}
\title{Parses and exports tweets to a Mongo DB collection, adding fields useful
for their analysis, following the same convention as the python scripts
the lab uses to collect tweets.}
\usage{
tweetsToMongo(file.name = NULL, ns = NULL, host = "localhost",
  username = "", password = "", verbose = TRUE)
}
\arguments{
\item{file.name}{string, name of the file where tweets were written.}

\item{ns}{string, namespace of the collection to which tweets will be added. Generally,
it will be of the form "database.collection". If the database or the collection do not exist,
they will be automatically created; if they exist, tweets will be appended.}

\item{host}{string host/port where mongo database is hosted. Default is localhost (127.0.0.1).}

\item{username}{string, username to be used for authentication purposes with MongoDB.}

\item{password}{string, password corresponding to the given username.}
}
\description{
\code{tweetsToMongo} read tweets in JSON format (downloaded e.g. using
the \code{filterStream} function in the \code{streamR} package), creates
an index and other fields that can facilitate their analysis, and stores
them in a MongoDB collection
}
\examples{
\dontrun{

## An example of an authenticated request using the ROAuth package,
## where consumerkey and consumer secret are fictitious.
## You can obtain your own at dev.twitter.com
  library(ROAuth)
  requestURL <- "https://api.twitter.com/oauth/request_token"
  accessURL <- "http://api.twitter.com/oauth/access_token"
  authURL <- "http://api.twitter.com/oauth/authorize"
  consumerKey <- "xxxxxyyyyyzzzzzz"
  consumerSecret <- "xxxxxxyyyyyzzzzzzz111111222222"
  my_oauth <- OAuthFactory$new(consumerKey=consumerKey,
    consumerSecret=consumerSecret, requestURL=requestURL,
    accessURL=accessURL, authURL=authURL)
  my_oauth$handshake(cainfo = system.file("CurlSSL", "cacert.pem", package = "RCurl"))

## capture 10 tweets mentioning the "Rstats" hashtag
  library(streamR)
  filterStream( file.name="tweets_rstats.json",
     track="rstats", tweets=10, oauth=my_oauth )

## exporting those 10 tweets to MongoDB (database name is "tweets" and
## collection name is "rstates")
  tweetsToMongo( file.name="tweets_rstats.json", ns="tweets.rstats",
      username="<USERNAME>", password="<PASSWORD>")

}
}
\author{
Pablo Barbera \email{pablo.barbera@nyu.edu}
}

