% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-geo.R
\name{getGeo}
\alias{getGeo}
\title{Returns geographic information about a location string}
\usage{
getGeo(location, verbose = FALSE,
  rdstk = "http://www.datasciencetoolkit.org",
  twofishes = "http://demo.twofishes.net/?autocomplete=true&maxInterpretations=1&")
}
\arguments{
\item{location}{location string for which information is desired}

\item{verbose}{If TRUE, provides additional information on console.}

\item{rdstk}{URL for data science toolkit}

\item{twofishes}{URL for twofishes geocoder}
}
\description{
\code{getGeo} connects to the Data Science Tool Kit and converts the
location string into a pair of coordinates, and then into information
(city, state, country, congressional district...) for those coordinates
}
\examples{
\dontrun{
## Download geographic information for "New York"
 getGeo("New York")
}
}
\author{
Pablo Barbera \email{pablo.barbera@nyu.edu}
}

