% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-timeline.R
\name{getTimeline}
\alias{getTimeline}
\title{Returns up to 3,200 recent tweets from a given user}
\usage{
getTimeline(filename, n = 3200, oauth_folder = "~/credentials",
  screen_name = NULL, id = NULL, since_id = NULL, trim_user = "true",
  sleep = 0.5, verbose = FALSE)
}
\arguments{
\item{filename}{file where tweets will be stored (in json format)}

\item{n}{number of tweets to be downloaded (maximum is 3,200)}

\item{oauth_folder}{folder where OAuth tokens are stored.}

\item{screen_name}{user name of the Twitter user for which his/her tweets
will be downloaded}

\item{id}{id of Twitter user for which his/her tweets will be downloaded
(Use either of these two arguments)}

\item{since_id}{id of the oldest tweet to be downloaded. Useful if, for
example, we're only interested in getting tweets sent after a certain
date.}

\item{trim_user}{if "true", downloaded tweets will include user object
embedded. If "false", only tweet information will be downloaded.}

\item{sleep}{numeric, number of seconds between API calls. Higher number
will increase reliability of API calls; lower number will increase speed.}

\item{verbose}{If TRUE, provides additional output in console about API
rate limits}
}
\description{
\code{getTimeline} connects to the REST API of Twitter and returns up to
3,200 recent tweets sent by these user. If the total number of tweets sent
by this user is less than 3,200 tweets, it will return all tweets.
}
\examples{
\dontrun{
## Download recent tweets by user "p_barbera"
 friends <- getTimeline(screen_name="p_barbera", oauth_folder="oauth")
}
}
\author{
Pablo Barbera \email{pablo.barbera@nyu.edu}
}

