% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get-users.R
\name{getUsers}
\alias{getUsers}
\title{Returns user data for up to 100 Twitter users}
\usage{
getUsers(oauth_folder = "~/credentials", screen_names = NULL, id = NULL,
  include_entities = "true", verbose = FALSE)
}
\arguments{
\item{oauth_folder}{folder where OAuth tokens are stored.}

\item{screen_names}{user names of the Twitter users}

\item{id}{ids of Twitter users}

\item{include_entities}{if "true", returned data will include most
recent tweet}

\item{verbose}{shows additional ouput about token usage in console}
}
\description{
\code{getUsers} connects to the REST API of Twitter and returns user
objects (user information) for up to 100 Twitter users, based on their
screen names or user IDs
}
\examples{
\dontrun{
## Download user data for user "p_barbera"
 userdata <- getUsers(screen_names="p_barbera", oauth_folder="oauth")
}
}
\author{
Pablo Barbera \email{pablo.barbera@nyu.edu}
}

