% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot-tweets.R
\name{plot.tweets}
\alias{plot.tweets}
\title{Plots number of tweets by hour or minute}
\usage{
\method{plot}{tweets}(tweets, breaks = "hours", return_plot = FALSE,
  missing_minutes = FALSE, hours = 0)
}
\arguments{
\item{tweets}{object of class tweets}

\item{breaks}{either "minutes" or "hours"}

\item{return_plot}{If \code{TRUE}, function returns plot as an object (useful
when saving to a file.)}

\item{missing_minutes}{If \code{TRUE}, minutes with zero tweets in the plot
where breaks is minutes will be interpolated.}

\item{hours}{numeric. When different from zero, will add that number of hours
to the timestamp of the tweets. That is, when hours is zero, then all times
are expressed in NY time.}
}
\description{
Plots number of tweets by hour or minute
}
\examples{
\dontrun{
## connect to the Mongo database
 mongo <- mongo.create("SMAPP_HOST:PORT", db="DATABASE")
 mongo.authenticate(mongo, username="USERNAME", password="PASSWORD", db="DATABASE")
 set <- "DATABASE.COLLECTION"

## extract all tweets that mention 'occupygezi' and do a quick plot
 tweets <- extract.tweets(set, string="occupygezi", fields="created_at")
 plot(tweets)
}
}
\author{
Pablo Barbera \email{pablo.barbera@nyu.edu}, Megan Metzger \email{megan.metzger@nyu.edu}
}

